last version

+++ app/Http/Controllers/Api/ProjectController.php
@@ -7,8 +7,10 @@ use App\Http\Requests\ProjectRequest;
 use App\Http\Resources\ProjectResource;
 use App\Models\Project;
 use App\Models\Picture;
+use App\Models\Technique;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 
 class ProjectController extends Controller
@@ -33,10 +35,15 @@ class ProjectController extends Controller
     {
 
 
-        $image = $request->image; //request()->file('image');
-        $imageName = $image->getClientOriginalName();
-        $imageName = time() . '_' . $imageName;
-        $image->move(public_path('/images'), $imageName);
+
+        if ($request->hasFile('image')) {
+            $image = $request->image;
+            $imageName = $image->getClientOriginalName();
+            $imageName = time() . '_' . $imageName;
+            $image->move(public_path('/images'), $imageName);
+        } else {
+            $imageName = 'default.jpg';
+        }
 
         $project = new Project();
         $project->image = $imageName;
@@ -46,18 +53,27 @@ class ProjectController extends Controller
         $project->project_url = $request->project_url;
         $project->category = $request->category;
         $project->save();
-
+        //uploading gallery
         if ($request->hasFile('files')) {
 
             foreach ($request->file('files') as $file) {
 
-                Log::info($file->getClientOriginalName());
                 $imageName = time() . '_' . $file->getClientOriginalName();
                 $file->move(public_path('/images'), $imageName);
 
                 Picture::create([
                     'project_id' => $project->id,
                     'picture' => $imageName,
+                    'created_at' => Carbon::now()
+                ]);
+            }
+        }
+        // creating techniques
+        if ($request->has('techniques')) {
+            foreach ($request->techniques as $technique) {
+                Technique::create([
+                    'project_id' => $project->id,
+                    'technique' => $technique,
                     'created_at' => Carbon::now()
                 ]);
             }
@@ -78,7 +94,8 @@ class ProjectController extends Controller
         $pictures = Picture::where('project_id', $Project->id)->get();
         $project
             = new ProjectResource($Project);
-        return ["project" => $project, "pictures" => $pictures];
+        $techniques = Technique::where('project_id', $Project->id)->get();
+        return ["project" => $project, "pictures" => $pictures, 'techniques' => $techniques];
     }
 
     /**
@@ -90,18 +107,32 @@ class ProjectController extends Controller
      */
     public function update(Request $request, Project $Project)
     {
+        Log::info($request);
 
         $path = public_path() . '/images/';
+        if ($request->new_image != 'null') { //code for remove old image
+            $file_old = $path . $request->image;
+            unlink($file_old);
 
-        //code for remove old image
-        $file_old = $path . $request->image;
-        unlink($file_old);
+            //code for add new image
+            $image = $request->new_image;
+            $imageName = $image->getClientOriginalName();
+            $imageName = time() . '_' . $imageName;
+            $image->move(public_path('/images'), $imageName);
+        } else {
+            $imageName = $request->image;
+        }
+        DB::table('techniques')->where('project_id', $Project->id)->delete();
 
-        //code for add new image
-        $image = $request->new_image;
-        $imageName = $image->getClientOriginalName();
-        $imageName = time() . '_' . $imageName;
-        $image->move(public_path('/images'), $imageName);
+        if ($request->has('techniques')) {
+            foreach ($request->techniques as $technique) {
+                Technique::create([
+                    'project_id' => $Project->id,
+                    'technique' => $technique,
+                    'created_at' => Carbon::now()
+                ]);
+            }
+        }
 
         $Project->update([
             'title' => $request->title,
